Added compute network to tpgtools and related utilities.
